requirements.txt
groq
sounddevice
soundfile
numpy
pynput
pyperclip
python-dotenv
pystray
pillow
